+++ src/Matchers/StringMatcher.hpp
@@ -15,13 +15,13 @@ class StringMatcher: public MatcherInterface
 
         Result* match(
             const vector<MatchableInterface *> &matchables,
-            int start,
+            size_t start,
             const forward_list<Result> &previousResults
         );
 
         Result* match(
             const vector<MatchableInterface *> &matchables,
-            int start
+            size_t start
         );
 
         Result* next();
